requirements.txt
opencv-python==4.10.0.84
tf_keras==2.18.0
deepface==0.0.93
pillow==11.1.0